Add logic to retrieve home screen info

+++ core/domain/src/main/java/com/igorvd/bitcoincharts/core/domain/formatter/BitcoinFormatter.kt
@@ -1,7 +1,7 @@
 package com.igorvd.bitcoincharts.core.domain.formatter
 
 interface BitcoinFormatter {
-    fun format(value: Float, fractionDigits: Int, appendUnit: Boolean = false): String
+    fun format(value: Double, fractionDigits: Int, appendUnit: Boolean = false): String
 
     companion object  {
         const val BITCOIN_UNIT = "BTC"
