+++ core/domain/src/main/java/com/igorvd/bitcoincharts/core/domain/formatter/BitcoinFormatterImpl.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 class BitcoinFormatterImpl @Inject constructor() : BitcoinFormatter {
 
     override fun format(
-        value: Float,
+        value: Double,
         fractionDigits: Int,
         appendUnit: Boolean
     ): String {
